refactor: move utility functions to sharedParser for better reusability
